requirements.txt
certifi==2017.4.17
chardet==3.0.3
cycler==0.10.0
decorator==4.0.11
dj-database-url==0.4.1
Django==1.11.1
django-extensions==1.7.9
django-html5==1.0.0
gunicorn==19.6.0
idna==2.5
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.3.0
matplotlib==2.0.2
nbformat==4.3.0
numpy==1.12.1
olefile==0.44
Pillow==4.1.1
plotly==2.0.9
psycopg2==2.6.2
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.17.3
six==1.10.0
traitlets==4.3.2
urllib3==1.21.1
whitenoise==3.2
